node: split up the test
